airline-booking/monolith-app/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
PyJWT==2.8.0
requests==2.31.0
Flask-CORS==4.0.0
boto3==1.34.0
python-dotenv==1.0.0
stripe==7.8.0

